backend/requirements.txt
fastapi
uvicorn[standard]
phidata
openai
supabase
python-dotenv
pydantic
pypdf
python-multipart